+++ src/components/Header/Header.tsx
@@ -1,21 +1,24 @@
+import { Link } from 'react-router-dom';
+import { AppRoute } from '../../const';
+
 const Header = () => (
   <header className="header">
     <div className="container">
       <div className="header__wrapper">
         <div className="header__left">
-          <a className="header__logo-link header__logo-link--active">
+          <Link to={AppRoute.Root} className="header__logo-link header__logo-link--active">
             <img className="header__logo" src="img/logo.svg" alt="6 cities logo" width="81" height="41"/>
-          </a>
+          </Link>
         </div>
         <nav className="header__nav">
           <ul className="header__nav-list">
             <li className="header__nav-item user">
-              <a className="header__nav-link header__nav-link--profile" href="#">
+              <Link className="header__nav-link header__nav-link--profile" to={AppRoute.Favoriets}>
                 <div className="header__avatar-wrapper user__avatar-wrapper">
                 </div>
                 <span className="header__user-name user__name">Oliver.conner@gmail.com</span>
                 <span className="header__favorite-count">3</span>
-              </a>
+              </Link>
             </li>
             <li className="header__nav-item">
               <a className="header__nav-link" href="#">
