+++ src/components/PlaceCardList/PlaceCardList.tsx
@@ -1,0 +1,51 @@
+import { Link } from 'react-router-dom';
+import { Offers } from '../../types';
+
+const PlaceCardList = ({offer, setId}: {offer : Offers; setId: (id: string) => void}) => {
+  const handleMouseEnter: React.MouseEventHandler = () => {
+    setId(offer.id);
+  };
+
+  const handleMouseLeave: React.MouseEventHandler = () => {
+    setId('');
+  };
+
+  return(
+    <article className="cities__card place-card" onMouseLeave={handleMouseLeave} onMouseEnter={handleMouseEnter}>
+      <div className="place-card__mark">
+        <span>Premium</span>
+      </div>
+      <div className="cities__image-wrapper place-card__image-wrapper">
+        <Link to={`/offer/${offer.id}`}>
+          <img className="place-card__image" src="img/apartment-01.jpg" width="260" height="200" alt="Place image"/>
+        </Link>
+      </div>
+      <div className="place-card__info">
+        <div className="place-card__price-wrapper">
+          <div className="place-card__price">
+            <b className="place-card__price-value">&euro;{offer.price}</b>
+            <span className="place-card__price-text">&#47;&nbsp;night</span>
+          </div>
+          <button className="place-card__bookmark-button button" type="button">
+            <svg className="place-card__bookmark-icon" width="18" height="19">
+              <use xlinkHref="#icon-bookmark"></use>
+            </svg>
+            <span className="visually-hidden">To bookmarks</span>
+          </button>
+        </div>
+        <div className="place-card__rating rating">
+          <div className="place-card__stars rating__stars">
+            <span style={{width: '80%'}}></span>
+            <span className="visually-hidden">Rating</span>
+          </div>
+        </div>
+        <h2 className="place-card__name">
+          <Link to={`/offer/${offer.id}`}>Beautiful &amp; luxurious apartment at great location</Link>
+        </h2>
+        <p className="place-card__type">Apartment</p>
+      </div>
+    </article>
+  );
+};
+
+export default PlaceCardList;
