+++ src/components/ReviewsForm/ReviewsForm.tsx
@@ -1,0 +1,89 @@
+import { useState } from 'react';
+import { StateForms } from '../Offer/OfferPage';
+
+const ReviewsForm = ({ setState }: {setState: React.Dispatch<React.SetStateAction<StateForms>>}) => {
+  const [text, setText] = useState('');
+  const [countStars,] = useState(5);
+
+  const handleSubmitReview: React.MouseEventHandler<HTMLButtonElement> = (e) => {
+    e.preventDefault();
+    setState((prevState) => ({
+      forms: [
+        ...prevState.forms,
+        {
+          // Как делать адекватные id
+          id: Math.random.toString(),
+          stars: countStars,
+          comment: text,
+        }
+      ]
+    }));
+    setText('');
+  };
+
+  return (
+    <form className="reviews__form form" action="#" method="post">
+      <label className="reviews__label form__label" htmlFor="review">Your review</label>
+      <div className="reviews__rating-form form__rating">
+        <input className="form__rating-input visually-hidden" name="rating" value="5" id="5-stars" type="radio"/>
+        <label htmlFor="5-stars" className="reviews__rating-label form__rating-label" title="perfect">
+          <svg className="form__star-image" width="37" height="33">
+            <use xlinkHref="#icon-star"></use>
+          </svg>
+        </label>
+
+        <input className="form__rating-input visually-hidden" name="rating" value="4" id="4-stars" type="radio"/>
+        <label htmlFor="4-stars" className="reviews__rating-label form__rating-label" title="good">
+          <svg className="form__star-image" width="37" height="33">
+            <use xlinkHref="#icon-star"></use>
+          </svg>
+        </label>
+
+        <input className="form__rating-input visually-hidden" name="rating" value="3" id="3-stars" type="radio"/>
+        <label htmlFor="3-stars" className="reviews__rating-label form__rating-label" title="not bad">
+          <svg className="form__star-image" width="37" height="33">
+            <use xlinkHref="#icon-star"></use>
+          </svg>
+        </label>
+
+        <input className="form__rating-input visually-hidden" name="rating" value="2" id="2-stars" type="radio"/>
+        <label htmlFor="2-stars" className="reviews__rating-label form__rating-label" title="badly">
+          <svg className="form__star-image" width="37" height="33">
+            <use xlinkHref="#icon-star"></use>
+          </svg>
+        </label>
+
+        <input className="form__rating-input visually-hidden" name="rating" value="1" id="1-star" type="radio"/>
+        <label htmlFor="1-star" className="reviews__rating-label form__rating-label" title="terribly">
+          <svg className="form__star-image" width="37" height="33">
+            <use xlinkHref="#icon-star"></use>
+          </svg>
+        </label>
+      </div>
+      <textarea
+        onChange={(e) => setText(e.target.value)}
+        value={text}
+        className="reviews__textarea form__textarea"
+        id="review"
+        name="review"
+        placeholder="Tell how was your stay, what you like and what can be improved"
+      >
+      </textarea>
+      <div className="reviews__button-wrapper">
+        <p className="reviews__help">
+          To submit review please make sure to set <span className="reviews__star">rating</span> and describe your stay with at least <b className="reviews__text-amount">50 characters</b>.
+        </p>
+        <button
+          onClick={(e) => handleSubmitReview(e)}
+          className="reviews__submit form__submit button"
+          type="submit"
+          disabled={text.length < 10}
+        >
+            Submit
+        </button>
+      </div>
+    </form>
+  );
+};
+
+export default ReviewsForm;
