--- src/components/privateRoute/private-route.tsx
@@ -1,18 +1,0 @@
-import { Navigate } from 'react-router-dom';
-import { AppRoute, AuthorizationStatus } from '../../const';
-
-type privateRouteProps = {
-  authorizationStatus: AuthorizationStatus;
-  children: JSX.Element;
-}
-
-const PrivateRoute = (props: privateRouteProps) => {
-  const {authorizationStatus, children} = props;
-  return (
-    authorizationStatus === AuthorizationStatus.Auth
-      ? children
-      : <Navigate to={AppRoute.Login}/>
-  );
-};
-
-export default PrivateRoute;
