+++ src/hooks.tsx
@@ -1,0 +1,3 @@
+import { AuthorizationStatus } from './const';
+
+export const useAuth = (authorizationStatus: AuthorizationStatus) => authorizationStatus === AuthorizationStatus.Auth;
