+++ src/index.tsx
@@ -1,14 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from './components/App/App';
-
-export type NumberOfOffers = {
-  countOffers: number;
-}
-
-const numberOfOffers: NumberOfOffers = {
-  countOffers: 6,
-};
+import { offers } from './mocks/offers';
 
 const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement
@@ -16,6 +9,6 @@ const root = ReactDOM.createRoot(
 
 root.render(
   <React.StrictMode>
-    <App {...numberOfOffers}/>
+    <App offers={offers}/>
   </React.StrictMode>
 );
